+++ Gemfile
@@ -11,7 +11,9 @@ gem "railties", RAILS_GEMS_VERSION
 
 gem "bootsnap", require: false
 gem "govuk_app_config"
-gem "govuk_message_queue_consumer"
+
+# Gems for publish_event_pipeline
+gem "govuk_message_queue_consumer", require: false
 
 group :test do
   gem "simplecov", require: false
