+++ lib/tasks/publish_event_pipeline.rake
@@ -1,4 +1,8 @@
-namespace :message_queue do
+require "govuk_message_queue_consumer"
+
+require "publish_event_pipeline/message_processor"
+
+namespace :publish_event_pipeline do
   desc "Create RabbitMQ queue for development environment"
   task create_queue: :environment do
     # The exchange, queue, and binding are created via Terraform outside of local development:
@@ -8,14 +12,14 @@ namespace :message_queue do
     bunny = Bunny.new
     channel = bunny.start.create_channel
     exch = Bunny::Exchange.new(channel, :topic, "published_documents")
-    channel.queue("search_api_v2_published_documents").bind(exch, routing_key: "*.*")
+    channel.queue(ENV.fetch("PUBLISH_EVENT_MESSAGE_QUEUE_NAME")).bind(exch, routing_key: "*.*")
   end
 
   desc "Listens to and processes messages from the published documents queue"
-  task consume_published_documents: :environment do
+  task process_messages: :environment do
     GovukMessageQueueConsumer::Consumer.new(
-      queue_name: "search_api_v2_published_documents",
-      processor: ConsumePublishedDocumentFromMessageQueue,
+      queue_name: ENV.fetch("PUBLISH_EVENT_MESSAGE_QUEUE_NAME"),
+      processor: PublishEventPipeline::MessageProcessor,
     ).run
   end
 end
