+++ spec/lib/publish_event_pipeline/publish_event_spec.rb
@@ -1,4 +1,6 @@
-RSpec.describe PublishEvent do
+require "publish_event_pipeline/publish_event"
+
+RSpec.describe PublishEventPipeline::PublishEvent do
   describe ".from_message_hash" do
     subject(:publish_event) { described_class.from_message_hash(message_hash) }
 
@@ -6,7 +8,7 @@ RSpec.describe PublishEvent do
       let(:message_hash) { json_fixture_as_hash("message_queue/message.json") }
 
       it "maps the message onto a PublishEvent" do
-        expected_document = Document.new(
+        expected_document = SearchableDocumentData.new(
           content_id: "f75d26a3-25a4-4c31-beea-a77cada4ce12",
           title: "Ebola medal for over 3000 heroes",
         )
